fix(k6): fix k6 run script to accept from proto
